Passionfruit 1.0.4
Changes in 1.0.4
-	Added feature allowing download parallelization by distributing the
image requests
across multiple hostnames. This will allow the speedup of loading,
especially for
galleries with 100+ images. It will only work if your site supports
wildcard subdomains.
*Note: This has only been tested on apache servers, see documentation
on enabling it.
-	Fixed cache bug forcing the main gallery html to be downloaded
instead of displayed
-	Replaced getimagesize() with exif_imagetype() for image verification
to speed up the script and to optimize memory use
-	Optimized the createthumbs script with the exif_imagetype function
-	Added more exif data support (doesn't do anything yet but is more
efficient then the last version)
-	Added hover and active effects for improved usability
-	Changed how meta.txt files work.
-	Added background-size: cover so folder/gallery background images will
fill the box